refactor: add shared tests
